refactor : game디렉토리로 move
