+++ src/main/java/baseball/domain/game/GameMessage.java
@@ -1,4 +1,4 @@
-package baseball.domain;
+package baseball.domain.game;
 
 public class GameMessage {
     public static final String GAME_START_MESSAGE = "숫자 야구 게임을 시작합니다.";
