+++ src/main/java/baseball/exception/game/GameResultTotalCountException.java
@@ -1,4 +1,4 @@
-package baseball.exception;
+package baseball.exception.game;
 
 public class GameResultTotalCountException extends IllegalArgumentException {
     private final static String GAME_RESULT_STRIKE_BALL_TOTAL_COUNT_ERROR_MESSAGE = "Strike와 Ball의 갯수는 3개를 넘을 수 없습니다.";
